#### What kind of change does this PR introduce?

e2e test speedup

- [X] PR title follows the guidelines defined in our [pull request documentation](https://github.com/ossf/scorecard/blob/main/CONTRIBUTING.md#pr-process)

#### What is the current behavior?
Theres 2 tests which are 1-2 orders of magnitude slower than other e2e tests:
- "Should return use of OSS-Fuzz"
  * `tensorflow/tensorflow` is a really big repo and this test downloads the repo.
- "Should attest to repos based on policy"
  * this one runs scorecard repeatedly on the same repos

#### What is the new behavior (if this is a feature change)?**
- "Should return use of OSS-Fuzz"
  * `ossf/scorecard-webapp` is a much smaller repo which is fuzzed by OSSFuzz
- "Should attest to repos based on policy"
  * split into 3 smaller tests
    * Since only one test dealt with `ossf-tests/scorecard-binauthz-test-good`, I changed the existing test to just test this one.
    * for the other two repos ( `ossf-tests/scorecard-binauthz-test-bad` and `ossf-tests/scorecard-attestor-code-review-e2e`), I changed the tests so they only fetch the results once, and just test the various policies against it. 

- [ ] Tests for the changes have been added (for bug fixes/features)

#### Which issue(s) this PR fixes
NONE
<!--
*Automatically closes linked issue when PR is merged.
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.

Fixes #

or

NONE
-->

#### Special notes for your reviewer

#### Does this PR introduce a user-facing change?

For user-facing changes, please add a concise, human-readable release note to
the `release-note`

(In particular, describe what changes users might need to make in their
application as a result of this pull request.)

<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required:
Enter your extended release note in the block below.
If the PR requires additional action from users switching to the new release,
include the string "ACTION REQUIRED".

For more information on release notes see: https://git.k8s.io/release/cmd/release-notes/README.md
-->

```release-note
NONE
```