@spencerschrock 2023-11-09
The speedups based on this PR should hopefully save 1-2 minutes total for the integration tests. The overall test time has a lot of variation between 6-10 minutes, but there's usually more stability in the individual test time.

Before:
Attestor test: 61s
OssFuzz test: 86s

After:
Attestor tests: 11 seconds total, (split into 1, 8, and 2s)
OssFuzz test: 7 seconds